+++ libs/cardano-ledger-binary/src/Cardano/Ledger/Binary/Decoding.hs
@@ -45,7 +45,7 @@ where
 
 import Cardano.Ledger.Binary.Decoding.Annotated
 import Cardano.Ledger.Binary.Decoding.DecCBOR
-import Cardano.Ledger.Binary.Decoding.Decoder
+import Cardano.Ledger.Binary.Decoding.Decoder hiding (getOriginalBytes)
 import Cardano.Ledger.Binary.Decoding.Drop
 import Cardano.Ledger.Binary.Decoding.Sharing
 import Cardano.Ledger.Binary.Decoding.Sized
@@ -143,8 +143,9 @@ deserialiseDecoder ::
   (forall s. Decoder s a) ->
   BSL.ByteString ->
   Either (Read.DeserialiseFailure, BS.ByteString) (a, BS.ByteString)
-deserialiseDecoder version decoder bs0 =
-  runST (supplyAllInput bs0 =<< Read.deserialiseIncremental (toPlainDecoder version decoder))
+deserialiseDecoder version decoder bsl =
+  runST $
+    supplyAllInput bsl =<< Read.deserialiseIncremental (toPlainDecoder (Just bsl) version decoder)
 {-# INLINE deserialiseDecoder #-}
 
 supplyAllInput ::